Split most of GUPPI driver class into a general base class.
